BSM-051 Remove aria-controls almost entirely
The aria-controls attribute is designed to help accessibility tools such
as screen readers to understand that interacting with the element will
do something to the other element(s) referenced in the aria-controls
attribute.

The aria-controls that this commit removes were not useful as they only
reference one element each, even though often multiple elements are
controlled by each "show/hide" link.

The aria-controls element allows multiple IDs to be listed, but to do so
is difficult due to the dynamic nature of the application.

Moreover, as the elements that are toggled are directly after the
element that does the toggling, we can simply rely on the (existing)
aria-expands attributes to implicitly convey the information we are
otherwise removing in this commit.